Support parameterized swift-testing tests
Swift-testing emits a `test` event at the beginning of a test run that
enumerates all the parameterized test cases to be executed.

This patch waits for the `test` event and then generates
`vscode.TestItem`s for each parameterized test execution, parenting them
to their test. Then when we recieve a `runStarted` event we enqueue all
the tests along with the newly created parameterized `TestItem`s.

Before a new test run starts we remove the existing parameterized
`TestItems` so we can regenerate them, as there may be a different
number of parameterized tests run with every execution.